remove uses of #if defined(DECRYPT_AT_CLIENT_SIDE)
replace #ifdef guards with plain if statements by defining a const bool
variable based on whether DECRYPT_AT_CLIENT_SIDE is defined.

the drawbacks of using #ifdef are that the code inside can't be syntax
checked unless the appropriate preprocessor directive is defined, and
that they are ugly and hard to follow when mixed with code.

since decrypt_at_client_side variable is a const, the compiler's dead
code elimination will be able to delete all the conditional code, so
this doesn't have any performance impact, while leading to nicer code.